+++ rsass/tests/spec/callable/mod.rs
@@ -1,0 +1,10 @@
+//! Tests auto-converted from "sass-spec/spec/callable"
+
+#[allow(unused)]
+fn runner() -> crate::TestRunner {
+    super::runner().with_cwd("callable")
+}
+
+mod arguments;
+
+mod parameters;