+++ rsass/tests/spec/non_conformant/errors/fn_varargs/at_start.rs
@@ -12,8 +12,8 @@ fn test() {
         "Error: expected \")\".\
          \n  ,\
          \n1 | @function test($rest...,$param) {}\
-         \n  |                        ^\
+         \n  |                         ^\
          \n  \'\
-         \n  input.scss 1:24  root stylesheet",
+         \n  input.scss 1:25  root stylesheet",
     );
 }